requirements.txt
attrs==21.4.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.0.0
python-dotenv==0.19.2
tomli==2.0.0